parsing directives and functional arrays
source files can now use the #use src.ml directive to load a file
as a prelude. parser uses new abstract syntax ELoadSrc as a
placeholder for this directive, and ELoadedSrc as a placeholder
to indicate the origin of an AST, for the ANF printer to use.

added Settings.djs_dir (set to $DJS_DIR) to use as a base for
all the files read/written. so the exe can now be run from any directory.

added Array type term and experimenting with functional array primitives.

added tests/ directory.