Object.prototype
Added this to djsPrelude.js. The only remaining part that needs to
be in tcDref.ml is creating the heap constraint for lObjectProto,
since its parent needs to be a special location lROOT. The rest of
the prototype can be defined (with externs) in DJS. So far, I've
only added Object.prototype.hasOwnProperty -- pretty cool that don't
need an inOwn operator like I always thought!