type/location parameter inference!
no more need for those pesky location parameters littered throughout!
or the type parameters for array operations. see the comments in
tcDref.ml for details. in short, some location parameters are inferred
by comparing the tuple value with the tuple type, and matching
tuple components of the form Ref L. location parameters that appear
as the parent link are substituted with the corresponding parent
location for a previously inferred location parameter. and type
parameters for Array A are inferred also by walking through heap
constraints.

now that these parameters can be omitted in both djs and djsLite
modes, i streamlined the desugaring of get/set/has because
separate cases are no longer needed (though burying the djsLite
desugaring in the more general case is a bit too subtle).