reviving while loops / tc-let-ann bug
adding the while loop stuff from September back in. so had
to add the mkLetRec back in, this time i'm not inserting all
the annotations on the inner lambdas that i used to.

bug in the TC-Let-Ann case. was doing

      let (n,g1) = tcAddBinding g h1 x s1 in
      tcExp g h1 goal e2;

instead of using g1 to check e2. i guess this rule wasn't being
fired before...