looping over arrays, assert tag(\x.e) = "Fun"
first examples of looping over arrays! the frame annotations are
currently hideous, and type checking is dog slow.

realized that {v != undefined} should be the default array invariant,
not Top, when it's omitted, since Top doesn't rule out undefined and
so the return value of getIdx is always possibly undefined.

was having trouble with (0 <= i < 2) implies (i = 0 \/ i = 1) through
the wrappers and everything (see tests/functional/numbers/la00.ml).
the problem was that i wasn't asserting that tag(\x.e) = "Fun" in the
theory, so couldn't rule out those values when trying to prove that
i belonged to the set of integers. all kinds of values should now
be non-overlapping based on tags after adding this missing case.