small opts from new THasTyp; BNstats; "Ref" tag
adding Neculas bNstats for profiling. not useful now since z3
running in separate process, and that constitutes the majority of
the running time.

reducing the number of calls to mustFlow in a couple ways
- setting cnf checks off by default
- removed a couple unnecessary calls to Sub.checkWorlds in tcExp
- changing THasTyp so that it keeps syntactic type terms "at the
  top level" so that they can be returned instead of extraction.
  so a new selfify function for variable cases keeps type terms
  manifest at the top level when possible.

    {v | v::U_1 /\ ... /\ v::U_n /\ p}

with these changes, the current test suite goes from
~11 minues to ~3 minutes.

need to add tag(v)="Ref" (and a new VRef logical constructor)
to strong references to ensure that they're truthy, so that the
better typeOf() can rule out the falsy case. adding this predicate
to strong reference types slows things way down though, several
times slower than the ~11 minutes before. so, not including this
by default for now...