arrays,thaw,refreeze
added a bunch of stuff for trying to compute a join invariant
rather than (not (= v undefined)) for an array tuple, before
realizing it's a better idea until VArray keeps the invariant
and other predicates separate.

also, starting to collect OWeak constraints to discharge, but
there's an issue with nested theV binders. so right now, not
checking the types of weak constraints as long as the location
names match. the typing rules don't manipulate those types
anyway...