recursive funcs
changed the JS parser and translation to EJS to hijack the
FuncStmtExpr (which gives a name to the function being defined)
for use in expression positions, and then using this form for
recursive definitions. (when used in stmt position, still treating
it as constructor functions.)

to typecheck fact00, needed to hack simpleJoinHeap, since each
branch had different constraints to be thrown away. quick and dirty
solution for now is to just return the heap var prefix if they're
the same, but need to handle this better in the future...