recursive functions on TreeNode.prototype!; apply
to avoid the difficulties of open recursion, instead of writing

  Looper.prototype.loop = function () { return 1 + this.loop() }

write

  Looper.prototype.loop = function loop( { return 1 + loop.apply(this) }

which is also valid JS, but gives the type system more traction since
it doesn't have to resolve the function being defined through the
prototype hierarchy (and the interaction between the weak objects
at the top level and the strong prototype chain will require something
new probably).

syntactically looking for f.apply(), which should be okay since
simple functions in DJS aren't desugared to objects anyway, so no
danger of f.apply being written in the dynamic execution.