half-baked marshaling of prelude environments
want to be able to run djs -marshalOutEnv to save the environments from
just processing the prelude, and then run djs -marshalInEnv foo.js to
load in the marshaled prelude. need to marshal/unmarshal other state
too (like idTypTerms), but not doing this now since seems like the payoff
is not so big (processing the prelude as usual takes about 0.5s (without typed
desugaring still) and unmarshaling the saved state takes about 0.25s).