towards smarter desugaring with JsProto and JsCtor
idea is to thread the boilerplate heap constraints for constructor
functions, and the monotonically increasing Foo.prototype objects,
with the -monotonicHeaps flags. not ready for primetime yet, since
will have to switch the macros from a pre-pass, to more of a delayed
expansion during desugaring, so that abbreviations can have the
shape of the current prototype object threaded through.