cleaning up addBinding/tcAddBinding
No longer using a flat list of identifiers curScope to check when shadowing
variables in the logic. Instead, keeping a stack curVars of lists to match
the logical contexts being pushed/popped. Removed removeBinding, since no
longer need to manually call something just to remove variables from the
scope list. To match, tcAddBinding no longer returns an integer that needs
to be passed to tcRemoveBindingN.

Also added checkBinder to let-bindings and lambdas to flag when a variable
is shadowed in the source program (as opposed to due to an issue with
type checking).