instantiating not (null :: Ref l); typeOf fast
want the above axiom for all strong locations, but axiomatizing
it requires defining a Box datatype in the logic instead of the
current int representation. instead, manually instantiating the
axiom along with the embedding of every type prediate w :: Ref(l).

in the future, would be better to instantiate once per type term.

typeOf is _way_ faster than in the paper.